+++ src/SpeedyWeather.jl
@@ -77,6 +77,10 @@ include("SpeedyTransforms/SpeedyTransforms.jl")
 using .SpeedyTransforms
 import .SpeedyTransforms: prettymemory
 
+# to be defined in GeoMakie extension
+export globe
+function globe end
+
 # Utility for GPU / KernelAbstractions
 include("gpu.jl")